+++ b2sdk/v2/__init__.py
@@ -22,9 +22,10 @@ from .session import B2Session
 from .sync import B2SyncPath
 from .transfer import DownloadManager, UploadManager
 
-# version & version utils
+# utils
 
 from .version_utils import rename_argument, rename_function
+from .utils import TempDir
 
 # raw_simulator
